requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
tomli==2.0.1
selenium==4.15.2
webdriver-manager==4.0.1
pylint==3.0.2
